fix test for new redigomock
The redigomock library fixed https://github.com/rafaeljusto/redigomock/issues/45, which
also fixed their behavior of not expecting a blank command when the redis connection is
returned to the pool. That broke my test since I had to tell redigomock to expect that
but now we no longer need to do that. This commit fixes my tests.